2/5/19
Linked javaScript. 
In javaScript I set it to strict mode, set the interval of the runclock to 1000, then set up a function stored the date and time as variables, then displayed the date and time as HTML, then calculated the days till January 1st, the calculated the hours left in the current day, then calculate the time left in the day in minutes and seconds, finally displayed and updated the time and the time left until January first.